+++ server/src/main/java/Method/Message.java
@@ -1,8 +1,8 @@
-package MethodUtil;
+package Method;
 
 import java.io.DataOutputStream;
 
-public class MessageUtil {
+public class Message {
 
     public static boolean sendMessage(DataOutputStream out, String fromuser, String touser, String body) {
         return true;
