+++ server/src/main/java/Method/Util/SendData.java
@@ -1,0 +1,10 @@
+package Method.Util;
+
+import java.io.DataOutputStream;
+
+public class SendData {
+
+    static boolean SendData(DataOutputStream out) {
+        return true;
+    }
+}
