+++ server/src/main/java/Object/Character.java
@@ -1,0 +1,4 @@
+package Object;
+
+public class Character {
+}
