+++ server/src/main/java/Object/ChatRoom.java
@@ -1,0 +1,4 @@
+package Object;
+
+public class ChatRoom {
+}
