+++ server/src/main/java/Service/ClientService.java
@@ -1,8 +1,5 @@
 package Service;
 
-import MethodUtil.GetUtil;
-import MethodUtil.MessageUtil;
-
 import java.io.*;
 import java.net.Socket;
 
@@ -75,17 +72,17 @@ public class ClientService implements Runnable {
             String determinator = elements[1];
             switch (determinator) {
                 case "Characters":
-                    assert GetUtil.getAndSendCharacters(out) : "getAndSendCharacters returns false";
+                    assert Method.Get.getAndSendCharacters(out) : "getAndSendCharacters returns false";
                     break;
                 case "Chatrooms":
-                    assert GetUtil.getAndSendChatrooms(out) : "getAndSendChatrooms returns false";
+                    assert Method.Get.getAndSendChatrooms(out) : "getAndSendChatrooms returns false";
                     break;
                 default:
                     System.err.printf("Bad Request: [%s] in Get method not available.", determinator);
             }
         } else if (method.equals("Message")) {
             assert elements.length == 4 : "Message method: Element length is not correct";
-            assert MessageUtil.sendMessage(out, elements[1], elements[2], elements[3]);
+            assert Method.Message.sendMessage(out, elements[1], elements[2], elements[3]);
         } else {
             System.err.printf("[%s] method not support.", method);
         }
